+++ test/concepts/wrapper_gtest.cpp
@@ -1,7 +1,7 @@
 #include <gtest_utils.h>
 
 #include <scl/feature/concepts/wrapper.h>
-#include <scl/feature/inplace/uninitialized.h>
+#include <scl/feature/executor/inplace/uninitialized.h>
 #include <scl/feature/wrapper.h>
 
 #include <string>