+++ test/type_traits/executor_gtest.cpp
@@ -1,7 +1,7 @@
 #include <gtest_utils.h>
 
-#include <scl/feature/inplace/plain.h>
-#include <scl/feature/inplace/uninitialized.h>
+#include <scl/feature/executor/inplace/plain.h>
+#include <scl/feature/executor/inplace/uninitialized.h>
 #include <scl/feature/type_traits/executor.h>
 #include <scl/utility/type_traits/forward_like.h>