+++ test/wrapper/wrapper_constructors_gtest.cpp
@@ -1,7 +1,7 @@
 #include <gtest_utils.h>
 
 #include <scl/feature/detail/wrapper_constructor_resolver.h>
-#include <scl/feature/inplace/plain.h>
+#include <scl/feature/executor/inplace/plain.h>
 #include <scl/feature/wrapper.h>
 
 #include <utility>